feat: make benchmark configurable
